Refectory: CSS 마진, Width, Height 제거 후 리펙토링 중
