+++ frontend/src/components/CustomModal/index.js
@@ -1,0 +1,11 @@
+import { FoodModal } from './FoodModal';
+import { OrderModal } from './OrderModal';
+import { EmailModal } from './EmailModal';
+
+const CustomModal = {
+  Food: FoodModal,
+  Order: OrderModal,
+  Email: EmailModal,
+};
+
+export { CustomModal };
