--- frontend/src/components/Modal/index.js
@@ -1,11 +1,0 @@
-import { FoodModal } from './FoodModal';
-import { OrderModal } from './OrderModal';
-import { EmailModal } from './EmailModal';
-
-const Modal = {
-  FoodModal,
-  OrderModal,
-  EmailModal,
-};
-
-export { Modal };
